refactor: move mockImportedToken
